refactor: Move repository classes to a dedicated package for better structure
